+++ app/widgets/AccountNext/AccountNext.php
@@ -1,5 +1,6 @@
 <?php
 
+use Movim\Librairies\XMPPtoForm;
 use Moxl\Xec\Action\Register\Set;
 
 use Movim\Session;
@@ -23,7 +24,7 @@ class AccountNext extends \Movim\Widget\Base
     {
         $form = $package->content;
 
-        $xtf = new \XMPPtoForm;
+        $xtf = new XMPPtoForm;
         $html = '';
         if (!empty($form->x)) {
             switch ($form->x->attributes()->xmlns) {